Put persistence into its own package.
1. Modularize sessions and puzzle storage into a separate package.
2. Make both the web and command-line clients use the new package.
3. Haven't started using database storage for puzzles (or anything
else).
