+++ cmd/susen-cli/main.go
@@ -20,27 +20,25 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"github.com/ancientHacker/susen.go/Godeps/_workspace/src/github.com/garyburd/redigo/redis"
 	"github.com/ancientHacker/susen.go/puzzle"
+	"github.com/ancientHacker/susen.go/storage"
 	"io"
 	"log"
 	"os"
 	"os/signal"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 )
 
 func main() {
-	// establish redis connection
-	redisInit()
-	if err := redisConnect(); err != nil {
+	// storage initialization
+	if err := storage.Connect(); err != nil {
+		log.Printf("Error during storage initialization: %v", err)
 		shutdown(startupFailureShutdown)
 	}
-	defer redisClose()
+	defer storage.Close()
 
 	// catch signals
 	shutdownOnSignal()
@@ -119,7 +117,7 @@ type commandInfo struct {
 	command     string
 	argInfo     string
 	description string
-	handler     func(*susenSession, *os.File, *request)
+	handler     func(*userSession, *os.File, *request)
 }
 
 // the command dispatch info is sorted for easy usage printing,
@@ -174,7 +172,7 @@ var (
 	showBindings = true
 )
 
-func markdownHandler(session *susenSession, w *os.File, r *request) {
+func markdownHandler(session *userSession, w *os.File, r *request) {
 	if len(r.args) > 0 {
 		switch r.args[0] {
 		case "on":
@@ -195,7 +193,7 @@ func markdownHandler(session *susenSession, w *os.File, r *request) {
 	}
 }
 
-func hintsHandler(session *susenSession, w *os.File, r *request) {
+func hintsHandler(session *userSession, w *os.File, r *request) {
 	if len(r.args) > 0 {
 		switch r.args[0] {
 		case "on":
@@ -216,12 +214,12 @@ func hintsHandler(session *susenSession, w *os.File, r *request) {
 	}
 }
 
-func backHandler(session *susenSession, w *os.File, r *request) {
-	session.removeStep()
+func backHandler(session *userSession, w *os.File, r *request) {
+	session.RemoveStep()
 	stateHandler(session, w, r)
 }
 
-func assignHandler(session *susenSession, w *os.File, r *request) {
+func assignHandler(session *userSession, w *os.File, r *request) {
 	var choice puzzle.Choice
 	var err error
 
@@ -232,17 +230,17 @@ func assignHandler(session *susenSession, w *os.File, r *request) {
 
 	// compute the index
 	idx := r.args[0]
-	if row := int(idx[0] - 'a'); row < 0 || row >= session.summary.SideLength {
+	if row := int(idx[0] - 'a'); row < 0 || row >= session.Summary.SideLength {
 		usageHandler(fmt.Sprintf("%s index (%s) row is out of range", r.command, idx), w, r)
 		return
 	} else if col, err := strconv.Atoi(idx[1:]); err != nil {
 		usageHandler(fmt.Sprintf("%s index (%s) column is not a number", r.command, idx), w, r)
 		return
-	} else if col < 1 || col > session.summary.SideLength {
+	} else if col < 1 || col > session.Summary.SideLength {
 		usageHandler(fmt.Sprintf("%s index (%s) column is out of range", r.command, idx), w, r)
 		return
 	} else {
-		choice.Index = (session.summary.SideLength * row) + col
+		choice.Index = (session.Summary.SideLength * row) + col
 	}
 
 	// read the value
@@ -252,11 +250,11 @@ func assignHandler(session *susenSession, w *os.File, r *request) {
 		return
 	}
 
-	update, e := session.puzzle.Assign(choice)
+	update, e := session.Puzzle.Assign(choice)
 	if e != nil {
 		fmt.Fprintf(w, "Assign failed: %v\n", e)
 	} else {
-		session.addStep()
+		session.AddStep()
 		if update.Errors != nil {
 			log.Printf("Assign to %s:%q gave errors; step %d is unsolvable.",
 				session.SID, session.PID, session.Step)
@@ -268,18 +266,18 @@ func assignHandler(session *susenSession, w *os.File, r *request) {
 	}
 }
 
-func stateHandler(session *susenSession, w *os.File, r *request) {
+func stateHandler(session *userSession, w *os.File, r *request) {
 	if useMarkdown {
-		fmt.Fprintf(w, "%s%s", session.puzzle.ValuesMarkdown(showBindings), session.puzzle.ErrorsMarkdown())
+		fmt.Fprintf(w, "%s%s", session.Puzzle.ValuesMarkdown(showBindings), session.Puzzle.ErrorsMarkdown())
 	} else {
-		fmt.Fprintf(w, "%s%s", session.puzzle.ValuesString(showBindings), session.puzzle.ErrorsString())
+		fmt.Fprintf(w, "%s%s", session.Puzzle.ValuesString(showBindings), session.Puzzle.ErrorsString())
 	}
 }
 
-func summaryHandler(session *susenSession, w *os.File, r *request) {
+func summaryHandler(session *userSession, w *os.File, r *request) {
 	fmt.Fprintf(w, "Session %q solving puzzle %q on solution step %d\n",
 		session.SID, session.PID, session.Step)
-	sum, err := session.puzzle.Summary()
+	sum, err := session.Puzzle.Summary()
 	if err != nil {
 		panic(err)
 	}
@@ -314,17 +312,8 @@ session handling
 
 */
 
-// A susenSession the user's current step in a puzzle solution.
-// Behind the scenes, we persist all the prior steps the user has
-// taken, so he can go back (undo) prior choices.
-type susenSession struct {
-	SID     string          // session ID
-	PID     string          // ID of puzzle being solved
-	Step    int             // current step
-	summary *puzzle.Summary // summary upon arriving at current step
-	puzzle  *puzzle.Puzzle  // puzzle for current step
-	Created string          // RFC3339 time when the session was created
-	Saved   string          // RFC3339 time when the session was last saved
+type userSession struct {
+	storage.Session
 }
 
 // cookie for the command line
@@ -356,7 +345,7 @@ func getCookie(w *os.File, r *request) string {
 }
 
 // sessionSelect: find or create the session for the current connection.
-func sessionSelect(w *os.File, r *request) *susenSession {
+func sessionSelect(w *os.File, r *request) *userSession {
 	id := getCookie(w, r)
 	// check to see if this is a force reset of the session
 	forceReset, resetID := r.command == "reset", ""
@@ -364,456 +353,21 @@ func sessionSelect(w *os.File, r *request) *susenSession {
 		resetID = r.args[0]
 	}
 	// create an in-memory session with this cookie
-	session := &susenSession{SID: id, Created: time.Now().Format(time.RFC3339)}
+	session := &userSession{storage.Session{SID: id, Created: time.Now().Format(time.RFC3339)}}
 	// load session from storage if possible, otherwise just initialize it
-	if session.redisLookup() {
+	if session.Lookup() {
 		log.Printf("Found session %v, puzzle %q, on step %d.", session.SID, session.PID, session.Step)
 		if forceReset {
-			session.startPuzzle(resetID)
+			session.StartPuzzle(resetID)
 		} else {
-			session.redisLoadStep()
+			session.LoadStep()
 		}
 	} else if forceReset {
-		session.startPuzzle(resetID)
+		session.StartPuzzle(resetID)
 	} else {
-		session.startPuzzle(defaultPuzzleID)
+		session.StartPuzzle(storage.DefaultPuzzleID())
 	}
 	return session
-}
-
-// startPuzzle: set the puzzle ID for the current session and
-// clear any existing solver steps for that puzzle ID.  If the
-// given puzzle ID is empty, try using the session's current
-// puzzle ID.  If the given puzzle ID is the special value
-// "default" (or unknown), use the default puzzle ID.
-func (session *susenSession) startPuzzle(pid string) {
-	// change to the given pid, making sure it's valid
-	if pid == "" {
-		pid = session.PID
-	} else if pid == "default" {
-		pid = defaultPuzzleID
-	}
-	session.summary = puzzleSummaries[pid]
-	if session.summary != nil {
-		session.PID = pid
-	} else {
-		session.PID, session.summary = defaultPuzzleID, puzzleSummaries[defaultPuzzleID]
-	}
-
-	// make the puzzle for the summary
-	p, e := puzzle.New(session.summary)
-	if e != nil {
-		log.Printf("Failed to create puzzle %q: %v", pid, e)
-		panic(e)
-	}
-	session.puzzle = p
-	session.redisStartPuzzle()
-	log.Printf("Reset session %v to start solving puzzle %q.", session.SID, session.PID)
-}
-
-// addStep: add a new current step with the current puzzle.
-func (session *susenSession) addStep() {
-	summary, err := session.puzzle.Summary()
-	if err != nil {
-		log.Printf("Failed to get summary of %s:%q step %d: %v", session.SID, session.PID, session.Step, err)
-		panic(err)
-	}
-	session.summary = summary
-	session.redisAddStep()
-	log.Printf("Added session %v:%v step %d.", session.SID, session.PID, session.Step)
-}
-
-// removeStep: remove the last step and restore the prior step's
-// puzzle.
-func (session *susenSession) removeStep() {
-	if session.Step > 1 {
-		session.redisRemoveStep()
-		log.Printf("Reverted session %v:%v to step %d.",
-			session.SID, session.PID, session.Step)
-	}
-}
-
-/*
-
-persistence layer
-
-*/
-
-// puzzle data
-var (
-	defaultPuzzleID = "standard-1"
-	puzzleSummaries = map[string]*puzzle.Summary{
-		"standard-1": &puzzle.Summary{
-			Geometry:   puzzle.StandardGeometryName,
-			SideLength: 9,
-			Values: []int{
-				4, 0, 0, 0, 0, 3, 5, 0, 2,
-				0, 0, 9, 5, 0, 6, 3, 4, 0,
-				0, 0, 0, 0, 0, 0, 0, 0, 8,
-				0, 0, 0, 0, 3, 4, 8, 6, 0,
-				0, 0, 4, 6, 0, 5, 2, 0, 0,
-				0, 2, 8, 7, 9, 0, 0, 0, 0,
-				9, 0, 0, 0, 0, 0, 0, 0, 0,
-				0, 8, 7, 3, 0, 2, 9, 0, 0,
-				5, 0, 2, 9, 0, 0, 0, 0, 6,
-			}},
-		"standard-2": &puzzle.Summary{
-			Geometry:   puzzle.StandardGeometryName,
-			SideLength: 9,
-			Values: []int{
-				0, 1, 0, 5, 0, 6, 0, 2, 0,
-				0, 0, 0, 0, 0, 3, 0, 1, 8,
-				0, 0, 0, 0, 7, 0, 0, 0, 6,
-				0, 0, 5, 0, 0, 0, 0, 3, 0,
-				0, 0, 8, 0, 9, 0, 7, 0, 0,
-				0, 6, 0, 0, 0, 0, 4, 0, 0,
-				5, 0, 0, 0, 4, 0, 0, 0, 0,
-				6, 4, 0, 2, 0, 0, 0, 0, 0,
-				0, 3, 0, 9, 0, 1, 0, 8, 0,
-			}},
-		"standard-3": &puzzle.Summary{
-			Geometry:   puzzle.StandardGeometryName,
-			SideLength: 9,
-			Values: []int{
-				9, 0, 0, 4, 5, 0, 0, 0, 8,
-				0, 2, 0, 0, 0, 0, 0, 0, 0,
-				0, 0, 0, 1, 7, 2, 4, 0, 0,
-				0, 7, 9, 0, 0, 0, 6, 8, 0,
-				2, 0, 0, 0, 0, 0, 0, 0, 5,
-				0, 4, 3, 0, 0, 0, 2, 7, 0,
-				0, 0, 8, 3, 2, 5, 0, 0, 0,
-				0, 0, 0, 0, 0, 0, 0, 6, 0,
-				4, 0, 0, 0, 1, 6, 0, 0, 3,
-			}},
-		"standard-4": &puzzle.Summary{
-			Geometry:   puzzle.StandardGeometryName,
-			SideLength: 9,
-			Values: []int{
-				9, 4, 8, 0, 5, 0, 2, 0, 0,
-				0, 0, 7, 8, 0, 3, 0, 0, 1,
-				0, 5, 0, 0, 7, 0, 0, 0, 0,
-				0, 7, 0, 0, 0, 0, 3, 0, 0,
-				2, 0, 0, 6, 0, 5, 0, 0, 4,
-				0, 0, 5, 0, 0, 0, 0, 9, 0,
-				0, 0, 0, 0, 6, 0, 0, 1, 0,
-				3, 0, 0, 5, 0, 9, 7, 0, 0,
-				0, 0, 6, 0, 1, 0, 4, 2, 3,
-			}},
-		"standard-5": &puzzle.Summary{
-			Geometry:   puzzle.StandardGeometryName,
-			SideLength: 9,
-			Values: []int{
-				0, 0, 0, 0, 0, 0, 0, 0, 0,
-				9, 0, 0, 5, 0, 7, 0, 3, 0,
-				0, 0, 0, 1, 0, 0, 6, 0, 7,
-				0, 4, 0, 0, 6, 0, 0, 8, 2,
-				6, 7, 0, 0, 0, 0, 0, 1, 3,
-				3, 8, 0, 0, 1, 0, 0, 9, 0,
-				7, 0, 5, 0, 0, 8, 0, 0, 0,
-				0, 2, 0, 3, 0, 9, 0, 0, 8,
-				0, 0, 0, 0, 0, 0, 0, 0, 0,
-			}},
-		"standard-6": &puzzle.Summary{
-			Geometry:   puzzle.StandardGeometryName,
-			SideLength: 9,
-			Values: []int{
-				2, 0, 0, 8, 0, 0, 0, 5, 0,
-				0, 8, 5, 0, 0, 0, 0, 0, 0,
-				0, 3, 6, 7, 5, 0, 0, 0, 1,
-				0, 0, 3, 0, 4, 0, 0, 9, 8,
-				0, 0, 0, 3, 0, 5, 0, 0, 0,
-				4, 1, 0, 0, 6, 0, 7, 0, 0,
-				5, 0, 0, 0, 0, 7, 1, 2, 0,
-				0, 0, 0, 0, 0, 0, 5, 6, 0,
-				0, 2, 0, 0, 0, 0, 0, 0, 4,
-			}},
-		"rectangular-1": &puzzle.Summary{
-			Geometry:   puzzle.RectangularGeometryName,
-			SideLength: 6,
-			Values: []int{
-				0, 4, 5, 1, 6, 0,
-				3, 0, 0, 0, 0, 0,
-				0, 5, 0, 6, 2, 1,
-				1, 0, 2, 3, 4, 0,
-				5, 0, 0, 2, 1, 6,
-				6, 0, 0, 0, 0, 0,
-			}},
-		"rectangular-2": &puzzle.Summary{
-			Geometry:   puzzle.RectangularGeometryName,
-			SideLength: 6,
-			Values: []int{
-				0, 0, 0, 2, 6, 0,
-				2, 0, 3, 0, 0, 0,
-				0, 5, 0, 0, 0, 6,
-				3, 2, 6, 0, 0, 1,
-				0, 0, 4, 0, 0, 0,
-				0, 0, 0, 5, 1, 4,
-			}},
-		"rectangular-3": &puzzle.Summary{
-			Geometry:   puzzle.RectangularGeometryName,
-			SideLength: 12,
-			Values: []int{
-				5, 7, 0, 6, 0, 0, 0, 0, 0, 1, 11, 12,
-				11, 0, 0, 0, 0, 0, 10, 0, 0, 0, 0, 3,
-				8, 0, 9, 0, 0, 0, 1, 0, 5, 7, 0, 0,
-				0, 0, 4, 2, 10, 11, 0, 0, 12, 0, 0, 8,
-				0, 0, 0, 0, 9, 6, 0, 1, 7, 0, 0, 0,
-				0, 9, 7, 0, 0, 0, 0, 2, 11, 0, 0, 0,
-				0, 0, 0, 8, 7, 0, 0, 0, 0, 11, 3, 0,
-				0, 0, 0, 11, 3, 0, 2, 5, 0, 0, 0, 0,
-				9, 0, 0, 3, 0, 0, 11, 8, 10, 6, 0, 0,
-				0, 0, 3, 7, 0, 10, 0, 0, 0, 12, 0, 2,
-				2, 0, 0, 0, 0, 3, 0, 0, 0, 0, 0, 11,
-				6, 11, 12, 0, 0, 0, 0, 0, 3, 0, 9, 4,
-			}},
-		"rectangular-4": &puzzle.Summary{
-			Geometry:   puzzle.RectangularGeometryName,
-			SideLength: 12,
-			Values: []int{
-				0, 11, 3, 0, 0, 0, 0, 0, 0, 6, 0, 0,
-				0, 7, 0, 0, 12, 0, 4, 0, 0, 3, 10, 8,
-				4, 6, 0, 0, 10, 11, 0, 0, 1, 0, 0, 7,
-				0, 0, 8, 9, 2, 0, 0, 0, 5, 0, 0, 0,
-				0, 0, 0, 0, 0, 9, 6, 0, 12, 8, 11, 0,
-				0, 5, 0, 0, 3, 0, 0, 11, 0, 9, 0, 0,
-				0, 0, 4, 0, 8, 0, 0, 9, 0, 0, 7, 0,
-				0, 9, 7, 3, 0, 10, 12, 0, 0, 0, 0, 0,
-				0, 0, 0, 11, 0, 0, 0, 1, 3, 12, 0, 0,
-				3, 0, 0, 7, 0, 0, 8, 2, 0, 0, 4, 1,
-				2, 8, 5, 0, 0, 12, 0, 4, 0, 0, 3, 0,
-				0, 0, 9, 0, 0, 0, 0, 0, 0, 7, 12, 0,
-			}},
-	}
-)
-
-// keys for persisted session values
-var (
-	sessionKeyFormat = "session:"
-	stepsKey         = "step:"
-)
-
-// current connection data
-var (
-	rdc     redis.Conn // open connection, if any
-	rdUrl   string     // URL for the open connection
-	rdEnv   string     // environment key prefix
-	rdMutex sync.Mutex // prevent concurrent connection use
-)
-
-// redisInit - look up redis info from the environment
-func redisInit() {
-	url := os.Getenv("REDISTOGO_URL")
-	db := os.Getenv("REDISTOGO_DB")
-	env := os.Getenv("REDISTOGO_ENV")
-	if db == "" {
-		db = "0" // default database
-	}
-	if url == "" {
-		rdUrl = "redis://localhost:6379/" + db
-	} else {
-		rdUrl = url + db
-	}
-	if env == "" {
-		if url == "" {
-			rdEnv = "local"
-		} else {
-			rdEnv = "dev"
-		}
-	} else {
-		rdEnv = env
-	}
-}
-
-// redisConnect: connect to the given Redis URL.  Returns the
-// connection, if successful, nil otherwise.
-func redisConnect() error {
-	conn, err := redis.DialURL(rdUrl)
-	if err == nil {
-		log.Printf("Connected to redis at %q (env: %q)", rdUrl, rdEnv)
-		rdc = conn
-		return nil
-	}
-	log.Printf("Can't connect to redis server at %q", rdUrl)
-	return err
-}
-
-// redisClose: close the given Redis connection.
-func redisClose() {
-	if rdc != nil {
-		rdc.Close()
-		log.Print("Closed connection to redis.")
-		rdc = nil
-	}
-}
-
-// redisExecute: execute the body with the redis connection.
-// Meant to be used inside a handler, because errors in execution
-// will panic back to the handler level.
-func redisExecute(body func() error) {
-	// wrap the body against runtime and database failures
-	wrapper := func() (err error) {
-		defer func() {
-			if r := recover(); r != nil {
-				if e, ok := r.(error); ok {
-					err = e
-				} else {
-					err = fmt.Errorf("%v", r)
-				}
-				log.Printf("Caught panic during redisExecute: %v", err)
-			}
-		}()
-		// Because redis connections can go away without warning,
-		// we ping to make sure the connection is alive, and try
-		// to reconnect if not.
-		if _, err := rdc.Do("PING"); err != nil {
-			log.Printf("PING failure with redis: %v", err)
-			redisClose()
-			err = redisConnect()
-			if err != nil {
-				log.Printf("Failed to reconnect to redis at %q", rdUrl)
-				return err
-			}
-		}
-		// connection is good; run the body
-		return body()
-	}
-	// grab the mutex and execute the body
-	rdMutex.Lock()
-	defer func(err error) {
-		rdMutex.Unlock()
-		if err != nil {
-			panic(err)
-		}
-	}(wrapper())
-}
-
-// redisKey - returns the session key
-func (session *susenSession) redisKey() string {
-	return rdEnv + ":SID:" + session.SID
-}
-
-// redisStepsKey - returns the key for the session's step array
-func (session *susenSession) redisStepsKey() string {
-	return session.redisKey() + ":Steps"
-}
-
-// redisLookup: lookup a session for an ID
-func (session *susenSession) redisLookup() (found bool) {
-	body := func() error {
-		vals, err := redis.Values(rdc.Do("HGETALL", session.redisKey()))
-		if len(vals) > 0 {
-			if err := redis.ScanStruct(vals, session); err != nil {
-				log.Printf("Redis error on parse of saved session %q: %v", session.SID, err)
-				return err
-			}
-			found = true
-			return nil
-		}
-		if err != nil {
-			log.Printf("Redis error on GET of session %q pid: %v", session.SID, err)
-			return err
-		}
-		log.Printf("No redis saved summary for session %q", session.SID)
-		return nil
-	}
-	redisExecute(body)
-	return
-}
-
-// redisLoadStep: load the current step from the saved summary
-func (session *susenSession) redisLoadStep() {
-	var bytes []byte
-	body := func() (err error) {
-		bytes, err = redis.Bytes(rdc.Do("LINDEX", session.redisStepsKey(), -1))
-		if err != nil {
-			log.Printf("Error on load of %s:%q step %d: %v", session.SID, session.PID, session.Step, err)
-		}
-		return
-	}
-	redisExecute(body)
-	session.redisUnmarshalStep(bytes)
-}
-
-// redisStartPuzzle: save a session that's just starting a puzzle
-func (session *susenSession) redisStartPuzzle() {
-	session.Saved = time.Now().Format(time.RFC3339)
-	session.Step = 1
-	bytes := session.redisMarshalStep()
-	body := func() (err error) {
-		rdc.Send("HMSET", redis.Args{}.Add(session.redisKey()).AddFlat(session)...)
-		rdc.Send("DEL", session.redisStepsKey())
-		_, err = rdc.Do("RPUSH", session.redisStepsKey(), bytes)
-		if err != nil {
-			log.Printf("Redis error on save of session %q after reset: %v", session.SID, err)
-		}
-		return
-	}
-	redisExecute(body)
-}
-
-// redisAddStep: add the current step to the saved summary
-func (session *susenSession) redisAddStep() {
-	session.Saved = time.Now().Format(time.RFC3339)
-	session.Step++
-	bytes := session.redisMarshalStep()
-	body := func() (err error) {
-		rdc.Send("HMSET", redis.Args{}.Add(session.redisKey()).AddFlat(session)...)
-		_, err = rdc.Do("RPUSH", session.redisStepsKey(), bytes)
-		if err != nil {
-			log.Printf("Redis error on save of %s:%q step %d: %v", session.SID, session.PID, session.Step, err)
-		}
-		return
-	}
-	redisExecute(body)
-}
-
-// redisRemoveStep: remove the last step from the saved session
-// and load the current step
-func (session *susenSession) redisRemoveStep() {
-	var bytes []byte
-	session.Saved = time.Now().Format(time.RFC3339)
-	session.Step--
-	session.summary = nil // free the current step's summary
-	body := func() (err error) {
-		rdc.Send("HMSET", redis.Args{}.Add(session.redisKey()).AddFlat(session)...)
-		rdc.Send("LTRIM", session.redisStepsKey(), 0, -2)
-		bytes, err = redis.Bytes(rdc.Do("LINDEX", session.redisStepsKey(), -1))
-		if err != nil {
-			log.Printf("Error on remove to %s:%q step %d: %v", session.SID, session.PID, session.Step, err)
-		}
-		return
-	}
-	redisExecute(body)
-	session.redisUnmarshalStep(bytes)
-}
-
-// redisMarshalStep - get JSON for the current step
-func (session *susenSession) redisMarshalStep() []byte {
-	bytes, err := json.Marshal(session.summary)
-	if err != nil {
-		log.Printf("Failed to marshal summary of %s:%q step %d (%+v) as JSON: %v",
-			session.SID, session.PID, session.Step, *session.summary, err)
-		panic(err)
-	}
-	return bytes
-}
-
-// redisUnmarshalStep - get puzzle for the saved step
-func (session *susenSession) redisUnmarshalStep(bytes []byte) {
-	var summary *puzzle.Summary
-	err := json.Unmarshal(bytes, &summary)
-	if err != nil {
-		log.Printf("Failed to unmarshal saved JSON of %s:%q step %d: %v",
-			session.SID, session.PID, session.Step, err)
-		panic(err)
-	}
-	session.summary = summary
-	session.puzzle, err = puzzle.New(session.summary)
-	if err != nil {
-		log.Printf("Failed to create puzzle for %s:%q step %d (%+v): %v",
-			session.SID, session.PID, session.Step, *session.summary, err)
-		panic(err)
-	}
 }
 
 /*
@@ -838,10 +392,7 @@ var alternateShutdown func(reason shutdownCause)
 
 // shutdown: process exit with logging.
 func shutdown(reason shutdownCause) {
-	// Get redis mutex and keep it to block other handlers from
-	// running.  Close the redis connection.
-	rdMutex.Lock()
-	redisClose()
+	storage.Close()
 
 	// for testing: run alternateShutdown instead, if defined
 	if alternateShutdown != nil {
